Update Ember tree

+++ net-misc/rsync-3.1.2/rsync-3.1.2.ebuild
@@ -1,6 +1,5 @@
-# Copyright 1999-2015 Gentoo Foundation
+# Copyright 1999-2017 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
-# $Id$
 
 EAPI="5"
 
@@ -14,9 +13,9 @@ SRC_URI="http://rsync.samba.org/ftp/rsync/src/${P}.tar.gz"
 LICENSE="GPL-3"
 SLOT="0"
 if [[ ${PV} != *_pre ]] ; then
-KEYWORDS="alpha amd64 arm arm64 hppa ia64 m68k ~mips ppc ppc64 s390 sh sparc x86 ~ppc-aix ~amd64-fbsd ~sparc-fbsd ~x86-fbsd ~x64-freebsd ~x86-freebsd ~hppa-hpux ~ia64-hpux ~x86-interix ~amd64-linux ~arm-linux ~ia64-linux ~x86-linux ~ppc-macos ~x64-macos ~x86-macos ~m68k-mint ~sparc-solaris ~sparc64-solaris ~x64-solaris ~x86-solaris"
+KEYWORDS="alpha amd64 arm arm64 hppa ia64 m68k ~mips ppc ppc64 s390 sh sparc x86 ~ppc-aix ~x64-cygwin ~amd64-fbsd ~sparc-fbsd ~x86-fbsd ~amd64-linux ~arm-linux ~x86-linux ~ppc-macos ~x64-macos ~x86-macos ~m68k-mint ~sparc-solaris ~sparc64-solaris ~x64-solaris ~x86-solaris"
 fi
-IUSE="acl iconv ipv6 static stunnel xattr"
+IUSE="acl examples iconv ipv6 static stunnel xattr"
 
 LIB_DEPEND="acl? ( virtual/acl[static-libs(+)] )
 	xattr? ( kernel_linux? ( sys-apps/attr[static-libs(+)] ) )
@@ -65,9 +64,11 @@ src_install() {
 	fi
 
 	# Install the useful contrib scripts
-	exeinto /usr/share/rsync
-	doexe support/*
-	rm -f "${ED}"/usr/share/rsync/{Makefile*,*.c}
+	if use examples ; then
+		exeinto /usr/share/rsync
+		doexe support/*
+		rm -f "${ED}"/usr/share/rsync/{Makefile*,*.c}
+	fi
 
 	eprefixify "${ED}"/etc/{,xinetd.d}/rsyncd*
